Full Stack Web Programming Project
This project was for one of my computer science courses at Yale. It connects with one of the Yale APIs, and generates a sql database of buildings and information. The website allows the user to create a profile, query the sql database for specific buildings, leave reviews, upload images, and upvote other user's comments.